refactor: add deserializer
