+++ src/stateful.ts
@@ -1,10 +1,12 @@
-import { bin2num, bsv } from '.';
-import { parseLiteral } from './internal';
-import { Int, Bytes, SupportedParamType, BaseType, Bool, PrivKey, PubKey, Sig, Ripemd160, Sha1, Sha256, SigHashType, SigHashPreimage, OpCodeType } from './scryptTypes';
+import { AbstractContract, Arguments, bin2num, bsv, num2bin, readBytes } from '.';
+import { deserializeArgfromHex } from './deserializer';
+import { flatternArg, parseLiteral } from './internal';
+import { Int, Bytes, SupportedParamType, ScryptType, Bool, PrivKey, PubKey, Sig, Ripemd160, Sha1, Sha256, SigHashType, SigHashPreimage, OpCodeType, TypeResolver } from './scryptTypes';
 
 export default class Stateful {
 
-
+  // state version
+  static readonly CURRENT_STATE_VERSION = 0n;
 
   static int2hex(n: Int): string {
     let asm = '';
@@ -67,7 +69,7 @@ export default class Stateful {
     } else if (typeof x === 'string') {
       const [value, type] = parseLiteral(x);
 
-      if (type === BaseType.PRIVKEY) {
+      if (type === ScryptType.PRIVKEY) {
         return Stateful.int2hex(value as bigint);
       }
 
@@ -94,7 +96,7 @@ export default class Stateful {
     } else if (typeof x === 'string') {
       const [value, type] = parseLiteral(x);
 
-      if (type === BaseType.PRIVKEY) {
+      if (type === ScryptType.PRIVKEY) {
         return Stateful.int2hex(value as bigint);
       }
 
@@ -105,36 +107,150 @@ export default class Stateful {
   }
 
 
+  /**
+ * only used for state contract
+ * @param args 
+ * @param isGenesis 
+ * @param finalTypeResolver 
+ * @returns 
+ */
+  static buildState(args: Arguments, isGenesis: boolean, resolver: TypeResolver): string {
+
+    const args_ = args.map(arg => {
+      return flatternArg(arg, resolver, { state: true, ignoreValue: false });
+    }).flat(Infinity) as Arguments;
+
+    if (args_.length <= 0) {
+      throw new Error('no state property found, buildContractState only used for state contract');
+    }
+
+    // append isGenesis which is a hidden built-in state
+    let state_hex = `${Stateful.toHex(isGenesis)}`;
+
+    state_hex += args_.map(a => Stateful.toHex(a.value)).join('');
+
+    //append meta
+    if (state_hex) {
+      const state_len = state_hex.length / 2;
+      state_hex += num2bin(BigInt(state_len), 4) + num2bin(Stateful.CURRENT_STATE_VERSION, 1);
+      return state_hex;
+    }
+
+    return state_hex;
+
+  }
+
+
+
+  static buildDefaultStateArgs(contract: AbstractContract): Arguments {
+
+    const dummyArgs = contract.stateProps.map(p => {
+      const dummyArg = Object.assign({}, p, { value: false });
+      return flatternArg(dummyArg, contract.resolver, { state: true, ignoreValue: true });
+    }).flat(Infinity) as Arguments;
+
+
+    const hexTemplateMap: Map<string, string> = new Map();
+
+    dummyArgs.forEach(p => {
+
+      if (p.type === ScryptType.INT || p.type === ScryptType.PRIVKEY) {
+        hexTemplateMap.set(`<${p.name}>`, Stateful.int2hex(0n));
+      } else if (p.type === ScryptType.BOOL) {
+        hexTemplateMap.set(`<${p.name}>`, Stateful.bool2hex(true));
+      } else if (p.type === ScryptType.BYTES
+        || p.type === ScryptType.PUBKEY
+        || p.type === ScryptType.SIG
+        || p.type === ScryptType.RIPEMD160
+        || p.type === ScryptType.SHA1
+        || p.type === ScryptType.SHA256
+        || p.type === ScryptType.SIGHASHTYPE
+        || p.type === ScryptType.SIGHASHPREIMAGE
+        || p.type === ScryptType.OPCODETYPE) {
+        hexTemplateMap.set(`<${p.name}>`, Stateful.bytes2hex('00'));
+      } else {
+        throw new Error(`param ${p.name} has unknown type ${p.type}`);
+      }
+    });
+
+    return contract.stateProps.map(param => deserializeArgfromHex(contract.resolver, Object.assign(param, {
+      value: undefined
+    }), hexTemplateMap, { state: true }));
+
+  }
+
+
+
   static deserializer(type: string, hex: string): SupportedParamType {
 
     switch (type) {
-      case BaseType.BOOL:
+      case ScryptType.BOOL:
         return Bool(Stateful.hex2bool(hex));
-      case BaseType.INT:
+      case ScryptType.INT:
         return Int(Stateful.hex2int(hex));
-      case BaseType.BYTES:
+      case ScryptType.BYTES:
         return Bytes(Stateful.hex2bytes(hex));
-      case BaseType.PRIVKEY:
+      case ScryptType.PRIVKEY:
         return PrivKey(Stateful.hex2int(hex));
-      case BaseType.PUBKEY:
+      case ScryptType.PUBKEY:
         return PubKey(Stateful.hex2bytes(hex));
-      case BaseType.SIG:
+      case ScryptType.SIG:
         return Sig(Stateful.hex2bytes(hex));
-      case BaseType.RIPEMD160:
+      case ScryptType.RIPEMD160:
         return Ripemd160(Stateful.hex2bytes(hex));
-      case BaseType.SHA1:
+      case ScryptType.SHA1:
         return Sha1(Stateful.hex2bytes(hex));
-      case BaseType.SHA256:
+      case ScryptType.SHA256:
         return Sha256(Stateful.hex2bytes(hex));
-      case BaseType.SIGHASHTYPE:
+      case ScryptType.SIGHASHTYPE:
         return SigHashType(Number(Stateful.hex2int(hex)));
-      case BaseType.SIGHASHPREIMAGE:
+      case ScryptType.SIGHASHPREIMAGE:
         return SigHashPreimage(Stateful.hex2bytes(hex));
-      case BaseType.OPCODETYPE:
+      case ScryptType.OPCODETYPE:
         return OpCodeType(Stateful.hex2bytes(hex));
       default:
         throw new Error(`<${type}> cannot be cast to ScryptType, only sCrypt native types supported`);
     }
 
   }
+
+
+  static parseStateHex(contract: AbstractContract, scriptHex: string): [boolean, Arguments] {
+
+    const metaScript = scriptHex.substr(scriptHex.length - 10, 10);
+    const version = Number(bin2num(metaScript.substr(metaScript.length - 2, 2)));
+    const stateLen = Number(bin2num(metaScript.substr(0, 8)));
+
+
+    const stateHex = scriptHex.substr(scriptHex.length - 10 - stateLen * 2, stateLen * 2);
+
+    const br = new bsv.encoding.BufferReader(Buffer.from(stateHex, 'hex'));
+
+    const opcodenum = br.readUInt8();
+
+    const isGenesis = opcodenum == 1;
+
+    const stateTemplateArgs: Map<string, string> = new Map();
+
+
+    const dummyArgs = contract.stateProps.map(p => {
+      const dummyArg = Object.assign({}, p, { value: false });
+      return flatternArg(dummyArg, contract.resolver, { state: true, ignoreValue: true });
+    }).flat(Infinity) as Arguments;
+
+    dummyArgs.forEach((param) => {
+      if (param.type === ScryptType.BOOL) {
+        const opcodenum = br.readUInt8();
+        stateTemplateArgs.set(`<${param.name}>`, opcodenum === 1 ? '01' : '00');
+      } else {
+        const { data } = readBytes(br);
+        stateTemplateArgs.set(`<${param.name}>`, data ? bsv.Script.fromASM(data).toHex() : '');
+      }
+    });
+
+    return [isGenesis, contract.stateProps.map(param => deserializeArgfromHex(contract.resolver, Object.assign(param, {
+      value: false
+    }), stateTemplateArgs, { state: true }))];
+  }
+
 }
